shipit/api: parse version per product
In gecko 68 we started building Fennec using mozilla-esr68 and beta
numbers became something like 68.1b1. `FirefoxVersion` won't accept this
kind of versions. It would be better to use per product classes to parse
versions.
