+++ src/shipit/api/shipit_api/api.py
@@ -9,7 +9,10 @@ from flask import abort
 from flask import current_app
 from flask import jsonify
 from flask_login import current_user
+from mozilla_version.gecko import DeveditionVersion
+from mozilla_version.gecko import FennecVersion
 from mozilla_version.gecko import FirefoxVersion
+from mozilla_version.gecko import ThunderbirdVersion
 from sqlalchemy.orm.exc import NoResultFound
 from werkzeug.exceptions import BadRequest
 
@@ -22,6 +25,7 @@ from shipit_api.config import SCOPE_PREFIX
 from shipit_api.models import Phase
 from shipit_api.models import Release
 from shipit_api.models import Signoff
+from shipit_api.release import Product
 from shipit_api.tasks import ActionsJsonNotFound
 from shipit_api.tasks import UnsupportedFlavor
 from shipit_api.tasks import fetch_actions_json
@@ -29,6 +33,13 @@ from shipit_api.tasks import generate_action_hook
 from shipit_api.tasks import render_action_hook
 
 logger = get_logger(__name__)
+
+VERSION_CLASSES = {
+    Product.FIREFOX.value: FirefoxVersion,
+    Product.FENNEC.value: FennecVersion,
+    Product.THUNDERBIRD.value: ThunderbirdVersion,
+    Product.DEVEDITION.value: DeveditionVersion,
+}
 
 
 def good_version(release):
@@ -39,8 +50,11 @@ def good_version(release):
     Example versions that cannot be parsed:
     1.1, 1.1b1, 2.0.0.1
     '''
+    product = release['product']
+    if product not in VERSION_CLASSES:
+        raise ValueError(f'Product {product} versions are not supported')
     try:
-        FirefoxVersion.parse(release['version'])
+        VERSION_CLASSES[product].parse(release['version'])
         return True
     except ValueError:
         return False
@@ -121,7 +135,7 @@ def list_releases(product=None, branch=None, version=None, build_number=None,
     releases = [r.json for r in releases.all()]
     # filter out not parsable releases, like 1.1, 1.1b1, etc
     releases = filter(good_version, releases)
-    return sorted(releases, key=lambda r: FirefoxVersion.parse(r['version']))
+    return sorted(releases, key=lambda r: VERSION_CLASSES[r['product']].parse(r['version']))
 
 
 def get_release(name):
